+++ packages/components/molecules/f-media-element/test/component/f-media-element.component.spec.js
@@ -1,14 +1,11 @@
-const MediaElement = require('../../test-utils/component-objects/f-media-element.component');
-
-const mediaElement = new MediaElement();
+import MediaElement from '../../test-utils/component-objects/f-media-element.component';
 
 describe('f-media-element component tests', () => {
-    beforeEach(async () => {
-        await mediaElement.load();
-    });
-
     it('should display the f-media-element component', async () => {
+        // Act
+        await MediaElement.load();
+
         // Assert
-        await expect(await mediaElement.isComponentDisplayed()).toBe(true);
+        await expect(await MediaElement.isComponentDisplayed()).toBe(true);
     });
 });