+++ packages/components/molecules/f-restaurant-card/test/component/f-restaurant-card.component.spec.js
@@ -1,14 +1,11 @@
-const RestaurantCard = require('../../test-utils/component-objects/f-restaurant-card.component');
-
-const restaurantCard = new RestaurantCard();
+import RestaurantCard from '../../test-utils/component-objects/f-restaurant-card.component';
 
 describe('f-restaurant-card component tests', () => {
-    beforeEach(async () => {
-        await restaurantCard.load();
-    });
-
     it('should display the f-restaurant-card component', async () => {
+        // Act
+        await RestaurantCard.load();
+
         // Assert
-        await expect(await restaurantCard.isComponentDisplayed()).toBe(true);
+        await expect(await RestaurantCard.isComponentDisplayed()).toBe(true);
     });
 });