+++ packages/components/molecules/f-alert/CHANGELOG.md
@@ -3,6 +3,16 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+
+v6.1.0
+------------------------------
+*July 5, 2022*
+
+### Changed
+- Update f-wdio-utils to v1.1.0
+- Update component objects / specs to use ES6 syntax.
+
+
 v6.0.0
 ------------------------------
 *June 24, 2022*
@@ -10,9 +20,10 @@ v6.0.0
 ### Changed
 - devDependency version range to match peerDependencies.
 
+
 v5.0.1
 ------------------------------
-*Jun 20, 2022*
+*June 20, 2022*
 
 ### Fixed
 - Alert messages are now read out to accessible users when displayed
@@ -20,7 +31,7 @@ v5.0.1
 
 v5.0.0
 ------------------------------
-*Jun 17, 2022*
+*June 17, 2022*
 
 ### Changed
 - Update to `@use` and `@forward` SASS syntax
@@ -28,7 +39,7 @@ v5.0.0
 
 v4.4.1
 ------------------------------
-*Jun 9, 2022*
+*June 9, 2022*
 
 ### Changed
 - Bumped wdio version and fixed breaking changes.