+++ packages/components/molecules/f-restaurant-card/CHANGELOG.md
@@ -3,6 +3,15 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v1.1.0
+------------------------------
+*July 5, 2022*
+
+### Changed
+- Update f-wdio-utils to v1.1.0
+- Update component objects / specs to use ES6 syntax.
+
+
 v1.0.0
 -----------------------------
 *June 22, 2022*